Split solidart.dart into part files
